move secrets and fix tests
